feat(ui): Remove old releases codebase
