+++ src/sentry/static/sentry/app/views/releasesV2/utils/index.tsx
@@ -1,42 +1,6 @@
 import round from 'lodash/round';
 
-import localStorage from 'app/utils/localStorage';
-import ConfigStore from 'app/stores/configStore';
-import OrganizationStore from 'app/stores/organizationStore';
-import {trackAnalyticsEvent} from 'app/utils/analytics';
 import {stringifyQueryObject, QueryResults} from 'app/utils/tokenizeSearch';
-
-const RELEASES_VERSION_KEY = 'releases:version';
-
-export const switchReleasesVersion = (version: '1' | '2', orgId: string) => {
-  localStorage.setItem(RELEASES_VERSION_KEY, version);
-  const user = ConfigStore.get('user');
-  trackAnalyticsEvent({
-    eventKey: version === '1' ? 'releases_v2.opt_out' : 'releases_v2.opt_in',
-    eventName:
-      version === '1' ? 'ReleasesV2: Go to releases1' : 'ReleasesV2: Go to releases2',
-    organization_id: parseInt(orgId, 10),
-    user_id: parseInt(user.id, 10),
-  });
-  location.reload();
-};
-
-export const wantsLegacyReleases = () => {
-  const version = localStorage.getItem(RELEASES_VERSION_KEY);
-
-  return version === '1';
-};
-
-export const decideReleasesVersion = async hasNewReleases => {
-  const {organization} = OrganizationStore.get();
-
-  // we are forcing everyone to version 2 now, we will delete the codebase behind v1 in a week or so
-  if (wantsLegacyReleases()) {
-    switchReleasesVersion('2', organization?.id ?? '0');
-  }
-
-  return hasNewReleases(true);
-};
 
 export const roundDuration = (seconds: number) => {
   return round(seconds, seconds > 60 ? 0 : 3);
